+++ engine/src/resources/model.rs
@@ -27,13 +27,13 @@ impl ModelData {
         let loader = loader.lock();
         let vertex_sub_buffer = loader
             .vertex_buffer_allocator
-            .allocate_slice(data.vertices.clone().len() as _)?;
+            .allocate_slice(data.vertices().len() as _)?;
         let index_sub_buffer = loader
             .index_buffer_allocator
-            .allocate_slice(data.indices.clone().len() as _)?;
+            .allocate_slice(data.indices().len() as _)?;
 
-        vertex_sub_buffer.write()?.copy_from_slice(&data.vertices);
-        index_sub_buffer.write()?.copy_from_slice(&data.indices);
+        vertex_sub_buffer.write()?.copy_from_slice(data.vertices());
+        index_sub_buffer.write()?.copy_from_slice(data.indices());
 
         Ok(Self {
             vertex_sub_buffer,
@@ -43,20 +43,39 @@ impl ModelData {
     }
 
     /// Returns the index and vertex data of this object.
-    pub fn get_data(&self) -> &Data {
+    pub fn data(&self) -> &Data {
         &self.data
     }
 
     /// Returns the size of this model in number of indices.
-    pub fn get_size(&self) -> usize {
-        self.data.indices.len()
+    pub fn size(&self) -> usize {
+        self.data.indices().len()
     }
 
-    pub(crate) fn get_vertex_buffer(&self) -> Subbuffer<[Vertex]> {
+    /// Writes data into the model influencing the model data of every instance of this model.
+    ///
+    /// Returns an error if the data given is bigger than the size of this objects vertices and indices.
+    pub fn write(&self, data: Data) -> Result<()> {
+        if self.vertex_sub_buffer.len() < data.vertices().len() as u64
+            || self.index_sub_buffer.len() < data.indices().len() as u64
+        {
+            return Err(anyhow::Error::msg("Writing into a model requires the given data size to match with the current data size. Load a new model instead."));
+        }
+
+        self.vertex_sub_buffer
+            .write()?
+            .copy_from_slice(data.vertices());
+        self.index_sub_buffer
+            .write()?
+            .copy_from_slice(data.indices());
+        Ok(())
+    }
+
+    pub(crate) fn vertex_buffer(&self) -> Subbuffer<[Vertex]> {
         self.vertex_sub_buffer.clone()
     }
 
-    pub(crate) fn get_index_buffer(&self) -> Subbuffer<[u32]> {
+    pub(crate) fn index_buffer(&self) -> Subbuffer<[u32]> {
         self.index_sub_buffer.clone()
     }
 }